requirements.txt
# CrewAI and instrumentation
crewai
crewai-tools
#aws-opentelemetry-distro==0.10.0
#duckduckgo-search
boto3
#botocore
python-dotenv
dotenv

langfuse
openinference-instrumentation-crewai
openinference-instrumentation-litellm

chainlit
